Refactoring tout les bugs, problèmes sécu, et 2 a 3 cents badsmells
Florent Combet

+++ Esporter/src/com/esporter/client/vue/component/MenuButton.java
@@ -153,7 +153,7 @@ public class MenuButton extends JRadioButton{
 					g2 = (Graphics2D)g.create();
 	
 					g2.setRenderingHint(RenderingHints.KEY_RENDERING, RenderingHints.VALUE_RENDER_QUALITY);
-					GradientPaint gp = new GradientPaint(0, height, new Color(0,164,210,(int)(100*alphaFadeOut)), 0, height/4, MasterFrame.COLOR_MENU_BACKGROUND);
+					GradientPaint gp = new GradientPaint(0, height, new Color(0,164,210,(int)(100*alphaFadeOut)), 0, ((float)height)/4, MasterFrame.COLOR_MENU_BACKGROUND);
 					g2.setPaint(gp);
 					
 					g2.fillRect(0, height/4, width, height*3/4);
